corrected methods into qap and parser into flowshop

+++ qap/climblarge.h
@@ -4,11 +4,11 @@
 #include "methods.h"
 #include <algorithm>   
 
-int64_t large_first(int ** D, int ** F, vector<int> * sol);
+float large_first(int ** D, int ** F, vector<int> * sol);
 
-int64_t large_best(int** D, int ** F, vector<int> * sol);
+float large_best(int** D, int ** F, vector<int> * sol);
 
-int64_t large_worst(int** D, int ** F, vector<int> * sol);
+float large_worst(int** D, int ** F, vector<int> * sol);
 
 
 #endif
