+++ qap/climbnun.h
@@ -5,10 +5,10 @@
 #include <algorithm>   
 
 
-int64_t nun_first(int ** D, int ** F, vector<int> * sol);
+float nun_first(int ** D, int ** F, vector<int> * sol);
 
-int64_t nun_ME(int** D, int ** F, vector<int> * sol);
+float nun_ME(int** D, int ** F, vector<int> * sol);
 
-int64_t nun_best(int** D, int ** F, vector<int> * sol);
+float nun_best(int** D, int ** F, vector<int> * sol);
 
 #endif
